IMG_avif.c: fix 'Shift amount too large' warnings in 32 bit builds.

+++ IMG_avif.c
@@ -115,14 +115,14 @@ static SDL_bool ReadAVIFHeader(SDL_RWops *src, Uint8 **header_data, size_t *head
         }
         read += 8;
 
-        size = (((size_t)magic[8] << 56) |
-                ((size_t)magic[9] << 48) |
-                ((size_t)magic[10] << 40) |
-                ((size_t)magic[11] << 32) |
-                ((size_t)magic[12] << 24) |
-                ((size_t)magic[13] << 16) |
-                ((size_t)magic[14] << 8) |
-                ((size_t)magic[15] << 0));
+        size = (((Uint64)magic[8] << 56) |
+                ((Uint64)magic[9] << 48) |
+                ((Uint64)magic[10] << 40) |
+                ((Uint64)magic[11] << 32) |
+                ((Uint64)magic[12] << 24) |
+                ((Uint64)magic[13] << 16) |
+                ((Uint64)magic[14] << 8) |
+                ((Uint64)magic[15] << 0));
     }
 
     if (size > INT_MAX) {
